Move all Chrono related logic in chrono_adapters
