+++ CMakeLists.txt
@@ -153,7 +153,7 @@ set(SEAHOWL_ELASTO_HEADERS
     include/seahowl/elasto/rotor_elasto.h
     include/seahowl/elasto/tower_elasto.h
     include/seahowl/elasto/mooring_elasto.h
-    include/seahowl/elasto/utils_elasto.h
+    include/seahowl/elasto/chrono_adapters.h
 )
 source_group(elasto FILES ${SEAHOWL_ELASTO_SOURCES} ${SEAHOWL_ELASTO_HEADERS})
