+++ include/seahowl/core/tower.h
@@ -3,8 +3,8 @@
 #include <memory>
 #include <vector>
 
+#include <seahowl/elasto/elasto.h>
 #include <seahowl/elasto/tower_elasto.h>
-#include <seahowl/elasto/utils_elasto.h>
 #include <seahowl/aero/tower_aero.h>
 #include <seahowl/core/reference_point.h>
