+++ include/seahowl/elasto/chrono_adapters.h
@@ -17,87 +17,6 @@
 namespace seahowl {
 namespace elasto {
 
-class RigidBody {
-  public:
-    virtual void set_position(Vector3d position) = 0;
-    virtual void set_mass(double mass) = 0;
-    virtual void set_inertia_diagonal(Vector3d inertia) = 0;
-    virtual void set_rotation(Quaternion rotation) = 0;
-    virtual void reset_forces() = 0;
-    virtual void accumulate_torque(Vector3d torque, bool is_local) = 0;
-    virtual double get_mass() = 0;
-    virtual Vector3d get_position() const = 0;
-    virtual Vector3d get_velocity() const = 0;
-    virtual Vector3d get_acceleration() const = 0;
-    virtual Quaternion get_rotation() const = 0;
-    virtual Vector3d get_direction() const = 0;
-    virtual Vector3d get_rotational_velocity_local() const = 0;
-    virtual Vector3d get_rotational_acceleration_local() const = 0;
-    virtual Vector3d get_rotational_velocity_global() const = 0;
-    virtual Vector3d get_rotational_acceleration_global() const = 0;
-};
-
-class NodeFEA {
-  public:
-    virtual void set_position(Vector3d position) = 0;
-    virtual void set_rotation(Quaternion rotation) = 0;
-    virtual void set_load(Vector3d force) = 0;
-    virtual void set_torque(Vector3d torque) = 0;
-    virtual Vector3d get_position() const = 0;
-    virtual Vector3d get_velocity() const = 0;
-    virtual Vector3d get_acceleration() const = 0;
-    virtual Quaternion get_rotation() const = 0;
-    virtual Vector3d get_direction() const = 0;
-    virtual Vector3d get_rotational_velocity_local() const = 0;
-    virtual Vector3d get_rotational_acceleration_local() const = 0;
-    virtual Vector3d get_rotational_velocity_global() const = 0;
-    virtual Vector3d get_rotational_acceleration_global() const = 0;
-    virtual Vector3d get_load() const = 0;
-    virtual Vector3d get_torque() const = 0;
-};
-
-class ElementFEA {
-  public:
-    std::vector<std::shared_ptr<NodeFEA>> nodes0;
-
-    virtual void set_nodes(std::shared_ptr<NodeFEA> node1, std::shared_ptr<NodeFEA> node2) = 0;
-    virtual double get_mass() = 0;
-
-    virtual void evaluate_position_rotation(double eta, Vector3d& position, Quaternion& rotation) = 0;
-};
-
-class BladeElementFEA : public ElementFEA {
-  public:
-    virtual void set_prebend(const Quaternion& prebend) = 0;
-};
-
-class LinkFix {
-  public:
-    virtual void initialize(std::shared_ptr<RigidBody> body1, std::shared_ptr<RigidBody> body2) = 0;
-    virtual void initialize(std::shared_ptr<NodeFEA> node1, std::shared_ptr<RigidBody> body2) = 0;
-};
-
-class LinkRevolute {
-  public:
-    virtual void initialize(std::shared_ptr<RigidBody> body1, std::shared_ptr<RigidBody> body2) = 0;
-};
-
-class MeshElasto {
-  public:
-    virtual void add(std::shared_ptr<NodeFEA> node) = 0;
-    virtual void add(std::shared_ptr<ElementFEA> element) = 0;
-};
-
-class SystemElasto {
-  public:
-    virtual Vector3d get_gravitational_acceleration() const = 0;
-    virtual void set_gravitational_acceleration(Vector3d gravitational_acceleration) = 0;
-    virtual void add(std::shared_ptr<RigidBody> body) = 0;
-    virtual void add(std::shared_ptr<MeshElasto> mesh) = 0;
-    virtual void add(std::shared_ptr<LinkFix> link) = 0;
-    virtual void add(std::shared_ptr<LinkRevolute> link) = 0;
-};
-
 class RigidBodyChrono : public RigidBody {
   public:
     std::shared_ptr<chrono::ChBody> chobj;
